Add confirm before delete

+++ index.php
@@ -10,7 +10,7 @@
     <li><?= $data['nom'] ?> - <?= $data['date'] ?>
         <ul>
             <li>
-                 <a href="supprimer.php?id=<?= $data['id'] ?>" style="color:red;"> <ion-icon name="trash-outline"></ion-icon> </a>
+                <a href="supprimer.php?id=<?= $data['id'] ?>" style="color:red;" onclick="if(!confirm('Voulez-vous supprimer?')){return false;}"> <ion-icon name="trash-outline"></ion-icon> </a>
                  <a href="formulaire_modifier.php?id=<?= $data['id'] ?>" style="color:green;"> <ion-icon name="create-outline"></ion-icon> </a>
                  Partagez : 
                  <a href="https://wa.me/?&text=<?= $data['message'] ?>" target="_blank"> <ion-icon name="logo-whatsapp"></ion-icon> </a>
